demo/kagent/pyproject.toml
[project]
name = "demoagent"
version = "0.1.0"
description = "Kagent agent for Evidra benchmark harness"
dependencies = [
  "google-adk==1.25.1",
]
requires-python = ">=3.13"
